Add Postgres backend and nginx frontend
This change replaces SQLite with Postgres as the database backend and
adds nginx as a proxy for the app running by gunicorn using the wsgi
interface provided by Django. Supervisor is used to keep gunicorn up and
running.

Both the frontend and the backend are kept in separate containers based
on official images and linked together through the existing app
container.