Add simple pipeline building and testing app image
A simple declarative pipeline that builds an app image and runs Django
unit tests inside of it.

It assumes a postgres container is up and running the database is
required for tests to pass.

The change also adds plugins required for Pipeline to work to the list
of plugins to install in the Jenkins image.

The next step is to configure a safe store for DockerHub credentials.