Push image to DockerHub (with credentials in Vault)
This describes how to launch and configure HashiCorp Vault Docker
container which is then use to store one's DockerHub credentials. These
are needed to login and push the built and tested app image to public
registry.

The method used to provide Jenkins with Vault credentials using a
short-lived token generated on the host machine, although works OK,
needs rethinking.