Add Jenkins pipeline for building and publishing images
Adds a quite simple, yet powerful, Jenkins pipeline which:

* builds the app image from its Dockerfile
* runs app's integration tests and feeds the results to the JUnit
Jenkins plugin
* pushes final image to DockerHub using credentials stored in HashiCorp
Vault

Jenkins itself lives in a container launched from a custom image that
allows to bake in plugins and skip the initial wizard.

IMPORTANT
The Pipeline assumes that a PostgreSQL database container and a Vault
container are set up and running. It also needs to have a Vault token
generated, so that Jenkins can access Vault.